Rename general.py to utils.py
